그룹,알림 API 추가 및 개선

+++ src/main/java/com/example/service/domain/group/GroupController.java
@@ -25,7 +25,7 @@ public class GroupController {
     }
 
     @GetMapping("/api/groups/{groupId}")
-    public ResponseEntity<Group> getGroupDetail(@PathVariable("groupId") int groupId){
+    public ResponseEntity<Group> getGroupDetail(@PathVariable("groupId") Long groupId){
         return ResponseEntity.ok(groupService.getGroupDetail(groupId));
     }
